Last commit,
updated readme

fixed an important bug: used next town of edge instead of distance of edge when calculating distances. (.first instead of .second.first)

funny result: the program ran successfully in the test cases except the larger input (recently sent) when that bug that exist. Meaning, in the first 4 test cases using the next edge as the distance instead of the distance of each edge does result the same.